requirements.txt
turtle
canvasvg
cairosvg
Pillow
tkinter